Fixed some security flaws, began work on profile page
