+++ package.json
@@ -16,9 +16,9 @@
   "name": "acorn",
   "description": "Acorn is a piece of Open Source Venue Management Software being developed for The Fallen Log at Kitchen 17",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node app.js"
   },
   "repository": {
     "type": "git",
